\name{dataEmilie}
\alias{dataEmilie}
\docType{data}
\title{
data Emilie.
}
\description{
data.
}
\usage{data(dataEmilie)}
\format{
  A data frame with 1313 observations on the following 15 variables.
  \describe{
    \item{\code{tempssuividc10_1}}{a numeric vector}
    \item{\code{inddc10}}{a numeric vector}
    \item{\code{t1delai1}}{a numeric vector}
    \item{\code{t2delai1}}{a numeric vector}
    \item{\code{DEM1_10}}{a numeric vector}
    \item{\code{AGEENTRE}}{a numeric vector}
    \item{\code{sexe2}}{a numeric vector}
    \item{\code{CEP}}{a numeric vector}
    \item{\code{SPT1_1}}{a numeric vector}
    \item{\code{SPT1_2}}{a numeric vector}
    \item{\code{IADL4_1}}{a numeric vector}
    \item{\code{MMS_TW1}}{a numeric vector}
    \item{\code{ISA1_15}}{a numeric vector}
    \item{\code{BENTON1}}{a numeric vector}
    \item{\code{COD_W1}}{a numeric vector}
  }
}
\details{
data.
}


\examples{
data(dataEmilie)

}
\keyword{datasets}
