% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataEx.R
\docType{data}
\name{dataEx}
\alias{dataEx}
\title{Simulated dataset}
\format{A data frame with 2429 observations on the following 6 variables.
\describe{ \item{i}{subject identification number}
\item{t}{time of measurement} \item{X1}{binary covariate} 
\item{X2}{continous Gaussian standard variable} 
\item{X3}{continous Gaussian variable} 
\item{Y}{longitudinal outcome}}}
\description{
The data were simulated ...
}
\keyword{datasets}
