% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exemple.R
\name{gradLMM}
\alias{gradLMM}
\title{Gradient of the log-likelihood of a linear mixed model with random intercept}
\usage{
gradLMM(b, Y, X, ni)
}
\arguments{
\item{b}{numeric vector specifying the parameter's values in the
following order : first the fixed effects and then the standard
deviation of the random intercept and of the independent error}

\item{Y}{numeric vector including the dependent outcome vector ordered by subject}

\item{X}{numeric matrix including the covariates}

\item{ni}{interger vector giving the number of repeated measures for each subject}
}
\value{
a vector containing the gradient of the log-likelihood of the linear mixed model at point b
}
\description{
Gradient of the log-likelihood of a linear mixed model with random intercept
}
